+++ src/timer.rs
@@ -414,15 +414,21 @@ impl<'d> Drop for TimerDriver<'d> {
 unsafe impl<'d> Send for TimerDriver<'d> {}
 
 #[cfg(feature = "nightly")]
-impl<'d> embedded_hal_async::delay::DelayUs for TimerDriver<'d> {
+impl<'d> embedded_hal_async::delay::DelayNs for TimerDriver<'d> {
+    async fn delay_ns(&mut self, ns: u32) {
+        let counter = core::cmp::max((self.tick_hz() * ns as u64) / 1_000_000_000, 1);
+
+        self.delay(counter).await.unwrap();
+    }
+
     async fn delay_us(&mut self, us: u32) {
-        let counter = core::cmp::max((self.tick_hz() * us as u64) / 1000000, 1);
+        let counter = core::cmp::max((self.tick_hz() * us as u64) / 1_000_000, 1);
 
         self.delay(counter).await.unwrap();
     }
 
     async fn delay_ms(&mut self, ms: u32) {
-        let counter = core::cmp::max((self.tick_hz() * ms as u64) / 1000, 1);
+        let counter = core::cmp::max((self.tick_hz() * ms as u64) / 1_000, 1);
 
         self.delay(counter).await.unwrap();
     }